# api/openapi.yaml
openapi: 3.1.0
info:
  title: PVM Agent Sandbox API
  version: 1.0.0
  description: |
    Control-plane REST surface of PVM (`uml-container`), a User-Mode Linux
    agent-sandbox manager. Covers the `/api` surface (Bearer auth via
    `API_SECRET`), the E2B-compatible root surface (`X-API-KEY` or Bearer),
    unauthenticated observability endpoints (`/healthz`, `/version`),
    `/metrics`, and the envd Connect-JSON compatibility service on :49983.
    The envd WebSocket version service (:49982) speaks a 5-byte envelope
    frame codec (1 flags + 4 len BE + JSON; end-stream flag 0x02) and is not
    expressible as an OpenAPI path; see the `envd` tag descriptions below.
  license:
    name: Repo license
...
servers:
  - url: http://localhost:8080
    description: API server (agentpvm api)
  - url: http://localhost:3000
    description: WebUI server (agentpvm webui; same routes)
  - url: http://localhost:49983
    description: envd Connect-JSON compatibility (no auth per E2B semantics)
security:
  - bearerAuth: []
...
paths:
...
  /sandboxes:
    get:
      tags: [e2b]
      summary: E2B compat — list sandboxes
      security: [{ apiKeyAuth: [] }, { bearerAuth: [] }]
      responses:
        "200":
          description: Sandbox list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Sandbox" }
    post:
      tags: [e2b]
      summary: E2B compat — create a sandbox
      security: [{ apiKeyAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                template_id: { type: string }
                metadata: { type: object, additionalProperties: true }
                timeout: { type: integer, description: seconds }
      responses:
        "200":
          description: Created sandbox
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sandbox" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /sandboxes/{sandboxID}:
    delete:
      tags: [e2b]
      summary: E2B compat — delete a sandbox
      security: [{ apiKeyAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: sandboxID, in: path, required: true, schema: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/OkJson" }
        "404": { $ref: "#/components/responses/NotFound" }
  /sandboxes/{sandboxID}/refreshes:
    post:
      tags: [e2b]
      summary: E2B compat — extend a sandbox deadline
      security: [{ apiKeyAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: sandboxID, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [duration]
              properties:
                duration: { type: integer, description: seconds to add }
      responses:
        "200":
          description: New deadline
          content:
            application/json:
              schema:
                type: object
                properties:
                  sandboxID: { type: string }
                  expires_at: { type: string, format: date-time }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------------------------------------------------------- envd compat
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token; the token is the controller's API_SECRET.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: E2B SDK-style API key header; value is the API_SECRET.
...
  responses:
...
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    OkJson:
      description: JSON status payload
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Human-readable error text.
...
    Sandbox:
      type: object
      properties:
        sandboxID: { type: string }
        templateID: { type: string }
        clientID: { type: string }
        started_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time }